Moving towards modular build
I moved the matrix development and solving stages into their own modules, and added a new module for defining reactions. This allows the simulation to be applied to any number of reactions/mechanisms. Still need to work on C steps (quick fix now to allow EC and EC', but CE not viable). EE does not work. Also added more functionality to plot and added future modules for defining geometry and building GUI